requirements.txt
Cheetah==2.4.4
Flask==0.12.1
Flask-API==0.7.1
Flask-OAuthlib==0.9.3
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
PyYAML==3.10
SecretStorage==2.0.0
bcrypt==3.1.3
python-jose==1.3.2
redis==2.10.5
